ML/requirements.txt
# Core Deep Learning
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0

# Medical Imaging
SimpleITK>=2.3.0
nibabel>=5.1.0
pydicom>=2.4.0
scikit-image>=0.21.0

# Data & Config
pyyaml>=6.0
pandas>=2.0.0
h5py>=3.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
imageio>=2.31.0

# Training Utilities
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.65.0
einops>=0.7.0

# Metrics
monai>=1.3.0

# Testing
pytest>=7.4.0
